Moved SimpleConfig to core package
